Add ability to claim a cell
The functionality was added for handling the user clicking a cell. If
the cell is not already being controlled, the user's marker will be
placed on the cell that was clicked and control will shift to the other
player. If the player clicks on a cell that either they or the opposing
player already controls no action is taken.