Add user and computer markers to state
The markers for which player will be X and O has been added to state.
Eventually the player will be able to choose their marker. A
componentDid Update lifecycle was created which checks for when the
currPlayer state is the same as the marker assigned to the computer.
When this happens, the method for the computer to select its cell will
run.