Eliminate ability to play before clearBoard
There was a gap in the logic that allowed a player to continue to play
the game after a winner was declared. In order to eliminate this a
gameOver state was created with an initial value of false. If gameOver
is true the player cannot claim a cell. When a winner is determined
the state changes. It is reset back to false in clearBoard.